+++ metax/taxafunc_ploter/heatmap_plot.py
@@ -1,5 +1,9 @@
 
-from distinctipy import distinctipy
+# for scaling data
+import pandas as pd
+from sklearn.preprocessing import StandardScaler
+import numpy as np
+# for heatmap plot
 import seaborn as sns
 import matplotlib.pyplot as plt
 from .get_distinct_colors import GetDistinctColors
@@ -16,6 +20,7 @@ class HeatmapPlot:
     # EXAMPLE: plot_top_taxa_func_heatmap_of_test_res(df_anova, sw.func, 200, 'f', (30,30))
         # reset sns style
         sns.set_theme()
+        plt.style.use('default')
 
     def rename_taxa(self, df):
         first_index = df.index[0]
@@ -36,19 +41,23 @@ class HeatmapPlot:
                                         value_type:str = 'p', fig_size:tuple|None = None, pvalue:float = 0.05, 
                                          col_cluster:bool = True, row_cluster:bool = True,
                                         cmap:str|None = None, rename_taxa:bool = True, font_size:int = 10, title:str = '',
-                                        show_all_labels:tuple = (False, False), return_type:str = 'fig'):
+                                        show_all_labels:tuple = (False, False), return_type:str = 'fig', scale = None, scale_method:str = 'maxmin'):
 
         
 
         func_name = self.tfa.func_name
         dft = df.copy()
         dft.reset_index(inplace=True)
-        type_map = {'f': ('f-statistic', 'Spectral_r', 1),
-                    'p': ('P-value', 'Reds_r', None),
-                    't': ('t-statistic', 'hot_r', 1)}
+        
 
-        plot_type = type_map.get(value_type, None)[0]
-        if plot_type is None:
+        
+        # type_map: 1st: plot_type, 2nd: cmap
+        type_map = {'f': ('f-statistic', 'Spectral_r'),
+                    'p': ('P-value', 'Reds_r'),
+                    't': ('t-statistic', 'RdBu_r')}
+
+        plot_type = type_map.get(value_type, "None")[0]
+        if plot_type == "None":
             raise ValueError("type must be 'p' or 'f' or 't'")
         
         
@@ -59,9 +68,9 @@ class HeatmapPlot:
                 
 
         if cmap is None:
-            plot_type, cmap, scale = type_map.get(value_type, None)
+            plot_type, cmap = type_map.get(value_type, None)
         else:
-            plot_type, _, scale = type_map.get(value_type, None)
+            plot_type, _ = type_map.get(value_type, None)
 
 
 
@@ -83,14 +92,14 @@ class HeatmapPlot:
             print(f"Top [{top_number}] significant: Taxa ({df_top.shape[1]}), Functions ({df_top.shape[0]})")
             df_plot = df_top.fillna(1) if plot_type == 'P-value' else df_top.fillna(0)
             
-            sns.set_style("white")
+            df_plot = self.scale_data(df = df_plot, scale_by = scale, method = scale_method)
 
             if return_type == 'fig':
                 sns_params = {
                     "center": 0,
                     "cmap": cmap,
-                    "linewidths": 0.2,
-                    "linecolor": (0, 0, 0, 0.1), # last value is alpha value, 0 is transparent, 1 is opaque
+                    "linewidths": 0.5, 
+                    "linecolor": 'gray',
                     "dendrogram_ratio": (0.1, 0.2),
                     "figsize": fig_size,
                     "col_cluster": col_cluster,
@@ -110,9 +119,9 @@ class HeatmapPlot:
                 fig.ax_heatmap.set_xlabel("Taxa")
                 fig.ax_heatmap.set_ylabel("Functions")
                 if title == "":
-                    title = f"Significant Differences between groups in Taxa-Function (Sorted by {plot_type} Top {top_number})"
+                    title = f"Significant Differences between groups in Taxa-Function (Top {top_number} sorted by {plot_type} , scaled by {scale})"
                 else:
-                    title = f"{title} (Sorted by {plot_type} Top {top_number})"
+                    title = f"{title} (Top {top_number} sorted by {plot_type} , scaled by {scale})"
 
                 plt.suptitle(title, weight='bold')
 
@@ -130,8 +139,8 @@ class HeatmapPlot:
                 cbar = fig.ax_heatmap.collections[0].colorbar
                 cbar.set_label(plot_type, rotation=90, labelpad=1)
                 cbar.ax.yaxis.set_ticks_position('left')
-                cbar.ax.yaxis.set_label_position('left')
-
+                cbar.ax.yaxis.set_label_position('right')
+                
                 plt.subplots_adjust(left=0.06, bottom=0.35, right=0.5, top=0.96, wspace=0.01, hspace=0.01)
 
                 plt.tight_layout()
@@ -147,7 +156,6 @@ class HeatmapPlot:
                     "row_cluster": row_cluster,
                     "method": self.linkage_method,
                     "metric": self.distance_metric,
-                    "standard_scale": scale,
                     "mask": df_top.isnull(),
                 }
                 fig = sns.clustermap(df_plot, **sns_params)
@@ -179,20 +187,28 @@ class HeatmapPlot:
                                        fig_size:tuple|None = None, pvalue:float = 0.05, scale = None, 
                                        col_cluster:bool = True, row_cluster:bool = True,
                                        cmap:str|None = None, rename_taxa:bool = True, font_size:int = 10,
-                                       show_all_labels:tuple = (False, False), rename_sample:bool = True
+                                       show_all_labels:tuple = (False, False), rename_sample:bool = True,
+                                       sort_by:str = 'P-value', scale_method:str = 'maxmin'
                                        ):
 
         dft = df.copy()
+
 
         scale_map ={None: None,
                     'None': None,
                     'row': 0,
                     'column': 1}
         scale = scale_map.get(scale)
+  
+        scale_map ={None: None,
+                    'None': None,
+                    'row': 0,
+                    'column': 1}
+        scale = scale_map.get(scale)
 
-        type_map = {'f': ('f-statistic', 'Spectral_r'),
-                    'p': ('P-value', 'RdGy_r'), # Reds, hot_r were used before
-                    't': ('t-statistic', 'RdGy_r')}
+        type_map = {'f': ('f-statistic', 'Spectral_r' if scale_method == 'maxmin' else 'RdBu_r'),
+                    'p': ('P-value', 'RdGy_r' if scale_method == 'maxmin' else 'RdBu_r'),
+                    't': ('t-statistic', 'RdGy_r' if scale_method == 'maxmin' else 'RdBu_r')}
 
         if cmap is None:
             plot_type, cmap = type_map.get(value_type)
@@ -210,14 +226,24 @@ class HeatmapPlot:
         dft = dft[dft['P-value'] < pvalue]
 
 
-        if 'f-statistic' in dft.columns.tolist():
-            dft = dft.sort_values(by=['P-value', 'f-statistic'], ascending=[True, False])
-            mat = dft.head(top_number)
-            mat= mat.drop(['P-value', 'f-statistic' ], axis=1)
-        elif 't-statistic' in dft.columns.tolist():
-            dft = dft.sort_values(by=['P-value', 't-statistic'], ascending=[True, False])
-            mat = dft.head(top_number)
-            mat= mat.drop(['P-value', 't-statistic'], axis=1)
+        if 'f-statistic' in dft.columns:
+            sort_column = 'f-statistic' if sort_by == 'f-statistic' else 'P-value'
+            ascending = sort_by == 'P-value' # True if sort_by is 'P-value' else False
+            dft = dft.sort_values(by=[sort_column], ascending=ascending)
+            mat = dft.head(top_number).drop(['P-value', 'f-statistic'], axis=1)
+
+        elif 't-statistic' in dft.columns:
+            if sort_by == 't-statistic':
+                dft['abs_t-statistic'] = dft['t-statistic'].abs()
+                sort_column = 'abs_t-statistic'
+                ascending = False
+            else:
+                sort_column = 'P-value'
+                ascending = True
+            
+            dft = dft.sort_values(by=[sort_column], ascending=ascending)
+            mat = dft.head(top_number).drop(['P-value', 't-statistic', 'abs_t-statistic'], axis=1, errors='ignore')
+
         else:
             raise ValueError("No 'f-statistic' or 't-statistic' in the dataframe")
 
@@ -226,9 +252,6 @@ class HeatmapPlot:
         if len(mat.columns) < 2:
             col_cluster = False
             
-        meta_df = self.tfa.meta_df
-        meta_name = self.tfa.meta_name
-
 
         if fig_size is None:
             fig_size = (30,30)
@@ -244,6 +267,9 @@ class HeatmapPlot:
            
             if rename_taxa:
                 mat = self.rename_taxa(mat)
+                
+            mat = self.scale_data(df = mat, scale_by = scale, method = scale_method)
+                
             sns_params = {
                 "center": 0,
                 "cmap": cmap,
@@ -253,7 +279,6 @@ class HeatmapPlot:
                 "row_cluster": row_cluster,
                 "method": self.linkage_method,
                 "metric": self.distance_metric,
-                "standard_scale": scale,
                 "col_colors": color_list,
                 "xticklabels": True if show_all_labels[0] else "auto",
                 "yticklabels": True if show_all_labels[1] else "auto",
@@ -264,13 +289,12 @@ class HeatmapPlot:
             fig.ax_heatmap.set_xticklabels(fig.ax_heatmap.get_xmajorticklabels(), fontsize=font_size, rotation=90)
             fig.ax_heatmap.set_yticklabels(fig.ax_heatmap.get_ymajorticklabels(), fontsize=font_size, rotation=0)
             plt.suptitle(
-                f"The Heatmap of intensity sorted by {plot_type} of Significant differences between groups (top {top_number})",
-                weight='bold'
+                f"The Heatmap of intensity of Significant differences between groups (top {top_number} sorted by {sort_by.split('(')[0]}, scaled by {scale})"
             )
             cbar = fig.ax_heatmap.collections[0].colorbar
             cbar.set_label("Intensity", rotation=90, labelpad=1)
             cbar.ax.yaxis.set_ticks_position('left')
-            cbar.ax.yaxis.set_label_position('left')
+            cbar.ax.yaxis.set_label_position('right')
 
             plt.subplots_adjust(left=0.05, bottom=0.11, right=0.5, top=0.96, wspace=0.01, hspace=0.01)
             
@@ -286,12 +310,12 @@ class HeatmapPlot:
     # Plot basic heatmap of matrix with color bar
     # EXAMPLE: plot_heatmap(sw, mat=get_top_intensity_matrix_of_test_res(df=df_anova, df_type='anova', top_num=100), 
                 #  title = 'The heatmap of top 100 significant differences between groups in Taxa-Function', 
-                #  fig_size=(30,30), scale=0)
+                #  fig_size=(30,30), scale='row')
     def plot_basic_heatmap(self,  df, title = 'Heatmap',fig_size:tuple|None = None, 
                     scale = None, col_cluster:bool = True, row_cluster:bool = True, 
                     cmap:str|None = None, rename_taxa:bool = True, font_size:int = 10,
                     show_all_labels:tuple = (False, False), rename_sample:bool = True, plot_mean:bool = False,
-                    sub_meta: str = "None"
+                    sub_meta: str = "None", scale_method:str = 'maxmin'
                     ):
         '''
         sub_meta is higher plot_mean, if sub_meta provided, plot_mean is False
@@ -307,13 +331,9 @@ class HeatmapPlot:
         if len(df.columns) < 2:
             col_cluster = False
         
-        scale_map ={None: None,
-            'None': None,
-            'row': 0,
-            'column': 1}
-        scale = scale_map.get(scale)
 
         mat = df.copy()
+        mat = self.scale_data(df = mat, scale_by = scale, method = scale_method)
 
         # if index is Taxon, rename index
         if rename_taxa:
@@ -332,7 +352,7 @@ class HeatmapPlot:
         # if only one column, remove col_cluster, set scale to None
         if len(mat.columns) < 2:
             col_cluster = False
-            scale = None
+            # scale = None
             
             
         sns_params = {
@@ -347,7 +367,6 @@ class HeatmapPlot:
             "row_cluster": row_cluster,
             "method": self.linkage_method,
             "metric": self.distance_metric,
-            "standard_scale": scale,
             "col_colors": color_list if not plot_mean else None,
             "xticklabels": True if show_all_labels[0] else "auto",
             "yticklabels": True if show_all_labels[1] else "auto",
@@ -358,13 +377,14 @@ class HeatmapPlot:
 
         fig.ax_heatmap.set_xticklabels(fig.ax_heatmap.get_xmajorticklabels(), fontsize=font_size, rotation=90)
         fig.ax_heatmap.set_yticklabels(fig.ax_heatmap.get_ymajorticklabels(), fontsize=font_size, rotation=0)
+        title = f"{title} (scaled by {scale})"
         plt.suptitle(title, weight='bold')
         
         cbar = fig.ax_heatmap.collections[0].colorbar
         cbar.set_label('Intensity', rotation=90, labelpad=1)
         cbar.ax.yaxis.set_ticks_position('left')
-        cbar.ax.yaxis.set_label_position('left')
-
+        cbar.ax.yaxis.set_label_position('right')
+        
         plt.subplots_adjust(left=0.05, bottom=0.15, right=0.5, top=0.96, wspace=0.01, hspace=0.01)
         plt.tight_layout()
         plt.show()
@@ -374,9 +394,14 @@ class HeatmapPlot:
         # For taxa-func heatmap
     # get the top intensity matrix of taxa-func table
     def get_top_across_table(self, df, top_number:str|int = 100, value_type:str = 'p', pvalue:float = 0.05, 
-                             rename_taxa:bool = False, col_cluster:bool = True, row_cluster:bool = True):
+                             rename_taxa:bool = False, col_cluster:bool = True, row_cluster:bool = True, scale:str|None = None, scale_method:str = 'maxmin'):
+        '''
+        For  taxa-func table only
+        Return a table x, and y are taxa and functions, respectively, and the values are the p-value, f-statistic or t-statistic.
+        '''
         res = self.plot_top_taxa_func_heatmap_of_test_res(df=df, top_number=top_number, value_type=value_type, pvalue=pvalue, 
-                                                          col_cluster=col_cluster, row_cluster=row_cluster, rename_taxa=rename_taxa, return_type='table')
+                                                          col_cluster=col_cluster, row_cluster=row_cluster, rename_taxa=rename_taxa, return_type='table',
+                                                          scale=scale, scale_method=scale_method)
         return res
     
     # plot heatmap for all condtion results of DESeq2All or DunnettAll
@@ -386,7 +411,7 @@ class HeatmapPlot:
                                        show_all_labels:tuple = (False, False), 
                                        return_type:str = 'fig', res_df_type:str = 'deseq2',
                                        p_type:str = 'padj', three_levels_df_type: str = 'same_trends',
-                                       show_col_colors:bool = True, remove_zero_col:bool = True):
+                                       show_col_colors:bool = True, remove_zero_col:bool = True, scale_method:str = 'maxmin'):
         """
         Plot a heatmap of all condition results.
 
@@ -419,7 +444,6 @@ class HeatmapPlot:
             - ValueError: If an error occurs during plotting.
 
         """
-        import numpy as np
         # keep 4 decimal places
         pvalue = round(pvalue, 4)
         color_list = None
@@ -476,7 +500,7 @@ class HeatmapPlot:
                 dft = self.rename_taxa(dft)
             # scale the data
             if scale:
-                dft = self.scale_data(dft, scale)
+                dft = self.scale_data(df = dft, scale_by = scale, method = scale_method)
             
             if cmap is None:
                 cmap = sns.color_palette("vlag", as_cmap=True, n_colors=30)
@@ -524,8 +548,8 @@ class HeatmapPlot:
                 cbar.set_label("log2FC" if res_df_type == 'deseq2' else 't-statistic', 
                                rotation=90, labelpad=1)
                 cbar.ax.yaxis.set_ticks_position('left')
-                cbar.ax.yaxis.set_label_position('left')
-
+                cbar.ax.yaxis.set_label_position('right')
+                
                 plt.subplots_adjust(left=0.05, bottom=0.15, right=0.5, top=0.96, wspace=0.01, hspace=0.01)
 
                 plt.tight_layout()
@@ -568,11 +592,10 @@ class HeatmapPlot:
     def plot_heatmap_of_dunnett_test_res(self, df,  pvalue:float = 0.05,scale:str|None = None,
                                        fig_size:tuple|None = None, col_cluster:bool = True, row_cluster:bool = True,
                                        cmap:str|None = None, rename_taxa:bool = True, font_size:int = 10,
-                                       show_all_labels:tuple = (False, False),  show_col_colors:bool = False    
+                                       show_all_labels:tuple = (False, False),  show_col_colors:bool = False, scale_method:str = 'maxmin'
                                        ):
         #! 只画t-statistic的heatmap, 用p-value过滤
-        import pandas as pd
-        import numpy as np
+
         
         
         pvalue = round(pvalue, 5)
@@ -609,7 +632,7 @@ class HeatmapPlot:
 
             # scale the data
             if scale:
-                dft = self.scale_data(dft, scale)
+                dft = self.scale_data(df = dft, scale_by = scale, method = scale_method)
                 
             
             if cmap is None:
@@ -647,8 +670,8 @@ class HeatmapPlot:
             cbar = fig.ax_heatmap.collections[0].colorbar
             cbar.set_label('t-statistic', rotation=90, labelpad=1)
             cbar.ax.yaxis.set_ticks_position('left')
-            cbar.ax.yaxis.set_label_position('left')
-
+            cbar.ax.yaxis.set_label_position('right')
+            
             plt.subplots_adjust(left=0.05, bottom=0.15, right=0.5, top=0.96, wspace=0.01, hspace=0.01)
 
             plt.tight_layout()
@@ -661,31 +684,8 @@ class HeatmapPlot:
 
 
     def get_heatmap_table_of_dunnett_res(self, df,  pvalue:float = 0.05,scale:str|None = None,
-                                        col_cluster:bool = True, row_cluster:bool = True, rename_taxa:bool = True):
-        import pandas as pd
-        import numpy as np
-        
-        def scale_data(dft, scale):
-            if scale == 'row':
-                # 对每行单独应用双向缩放
-                for index, row in dft.iterrows():
-                    max_val = abs(row).max()
-                    if max_val != 0:
-                        dft.loc[index] = row / max_val
-            elif scale == 'col':
-                # 对每列单独应用双向缩放
-                for col in dft:
-                    max_val = abs(dft[col]).max()
-                    if max_val != 0:
-                        dft[col] = dft[col] / max_val
-            elif scale == 'all':
-                # 对整个数据框应用双向缩放
-                max_val = abs(dft.values).max()
-                if max_val != 0:
-                    dft = dft / max_val
+                                        col_cluster:bool = True, row_cluster:bool = True, rename_taxa:bool = True, scale_method:str = 'maxmin'):
 
-            return dft
-        
         
         dft = self.tfa.CrossTest.extrcat_significant_stat_from_dunnett(df, p_value=pvalue)
 
@@ -715,7 +715,7 @@ class HeatmapPlot:
 
             # scale the data
             if scale:
-                dft = scale_data(dft, scale)
+                dft = self.scale_data(df = dft, scale_by = scale, method = scale_method)
                 
             
             from matplotlib.colors import TwoSlopeNorm
@@ -754,14 +754,15 @@ class HeatmapPlot:
 
     def get_top_across_table_basic(self, df, top_number:int = 100, value_type:str = 'p', 
                                        fig_size:tuple|None = None, pvalue:float = 0.05, scale = None, 
-                                       col_cluster:bool = True, row_cluster:bool = True, cmap:str|None = None, rename_taxa:bool = False):
+                                       col_cluster:bool = True, row_cluster:bool = True, cmap:str|None = None, rename_taxa:bool = False, 
+                                       scale_method:str = 'maxmin'):
+        '''
+        for the tables of taxa, func, protein, peptides and custom table
+        Return a intensity matrix of top significant results.
+        '''
         dft = df.copy()
+        
 
-        scale_map ={None: None,
-                    'None': None,
-                    'row': 0,
-                    'column': 1}
-        scale = scale_map.get(scale)
 
         type_map = {'f': ('f-statistic', 'Spectral_r'),
                     'p': ('P-value', 'Reds_r'),
@@ -788,6 +789,7 @@ class HeatmapPlot:
             mat = dft.head(top_number)
             mat= mat.drop(['P-value', 'f-statistic' ], axis=1)
         elif 't-statistic' in dft.columns.tolist():
+            #! t-statistic incude positive and negative value
             dft = dft.sort_values(by=['P-value', 't-statistic'], ascending=[True, False])
             mat = dft.head(top_number)
             mat= mat.drop(['P-value', 't-statistic'], axis=1)
@@ -797,8 +799,10 @@ class HeatmapPlot:
         if len(mat.columns) < 2:
             col_cluster = False
             
-        meta_df = self.tfa.meta_df
-        meta_name = self.tfa.meta_name
+        mat = self.scale_data(df = mat, scale_by = scale, method = scale_method)
+            
+        # meta_df = self.tfa.meta_df
+        # meta_name = self.tfa.meta_name
 
 
         if fig_size is None:
@@ -812,8 +816,10 @@ class HeatmapPlot:
             groups_list = []
             new_col_names = []
             for i in col_names:
-                group = meta_df[meta_df['Sample'] == i]
-                group = group[meta_name].values[0]
+                # group = meta_df[meta_df['Sample'] == i]
+                group = self.tfa.meta_df[self.tfa.meta_df['Sample'] == i]
+                # group = group[meta_name].values[0]
+                group = group[self.tfa.meta_name].values[0]
                 new_col_names.append(f'{i} ({group})')
                 groups_list.append(group)
             color_list = self.assign_colors(groups_list)
@@ -831,7 +837,6 @@ class HeatmapPlot:
                 "row_cluster": row_cluster,
                 "method": self.linkage_method,
                 "metric": self.distance_metric,
-                "standard_scale": scale,
                 "col_colors": color_list,
             }
 
@@ -858,27 +863,83 @@ class HeatmapPlot:
         #     plt.close('all')
 
 
-    def scale_data(self, dft, scale):
-        try:
-            if scale == 'row':
-                # 对每行单独应用双向缩放
-                for index, row in dft.iterrows():
-                    max_val = abs(row).max()
-                    if max_val != 0:
-                        dft.loc[index] = row / max_val
-            elif scale == 'col':
-                # 对每列单独应用双向缩放
-                for col in dft:
-                    max_val = abs(dft[col]).max()
-                    if max_val != 0:
-                        dft[col] = dft[col] / max_val
-            elif scale == 'all':
-                # 对整个数据框应用双向缩放
-                max_val = abs(dft.values).max()
-                if max_val != 0:
-                    dft = dft / max_val
-                    
-        except Exception as e:
-            print(f'Error: {e}')
 
-        return dft
+
+    def scale_data(self, df: pd.DataFrame, scale_by: str|None = None, method: str|None = 'maxmin') -> pd.DataFrame:
+        scale_by = scale_by.lower() if scale_by else 'all'
+        method = method.lower() if method else 'maxmin'
+        
+        print(f"Scaling the data by [{scale_by}] using method [{method}]")
+
+        try:
+            if method == 'zscore':
+                scaler = StandardScaler()
+
+                if scale_by == 'row':
+                    if (df == 0).all(axis=1).any():
+                        raise ValueError("One or more rows are entirely zero, cannot perform z-score scaling.")
+                    
+                    df = df.apply(lambda row: pd.Series(scaler.fit_transform(row.values.reshape(-1, 1)).flatten(), index=row.index), axis=1)
+
+                elif scale_by == 'col':
+                    if (df == 0).all(axis=0).any():
+                        raise ValueError("One or more columns are entirely zero, cannot perform z-score scaling.")
+                    
+                    df = df.apply(lambda col: pd.Series(scaler.fit_transform(col.values.reshape(-1, 1)).flatten(), index=col.index), axis=0)
+                
+                else:  # 'all'
+                    if np.all(df.values.flatten() == 0):
+                        raise ValueError("All data are zero, cannot perform z-score scaling.")
+                    
+                    df = pd.DataFrame(scaler.fit_transform(df), index=df.index, columns=df.columns)
+
+            else:  # 'maxmin'
+                if scale_by == 'row':
+                    max_val = df.abs().max(axis=1)
+                    df = pd.DataFrame([row / max_val.loc[index] if max_val.loc[index] != 0 else row for index, row in df.iterrows()], index=df.index, columns=df.columns)
+                elif scale_by == 'col':
+                    max_val = df.abs().max()
+                    for col in df.columns:
+                        if max_val[col] != 0:
+                            df[col] = df[col] / max_val[col]
+                else:  # 'all'
+                    max_val = df.abs().values.max()
+                    df = df / max_val if max_val != 0 else df
+
+        except Exception as e:
+            print(f'Error in scaling the data: {e}')
+            raise e
+        
+        return df
+
+
+
+    # def scale_data(self, df, scale:str|None = None):
+    #     ''''
+    #     Scale the data by Max-Min scaling method
+    #     '''
+    #     try:
+    #         scale = scale.lower() if scale else None
+            
+    #         print(f"Scaling the data by [{scale}]")
+    #         if scale == 'row':
+    #             for index, row in df.iterrows():
+    #                 max_val = abs(row).max()
+    #                 if max_val != 0:
+    #                     df.loc[index] = row / max_val
+    #         elif scale == 'col':
+    #             for col in df:
+    #                 max_val = abs(df[col]).max()
+    #                 if max_val != 0:
+    #                     df[col] = df[col] / max_val
+    #         elif scale == 'all':
+    #             max_val = abs(df.values).max()
+    #             if max_val != 0:
+    #                 df = df / max_val
+    #         else:
+    #             print('No scale applied')
+                    
+    #     except Exception as e:
+    #         print(f'Error in scaling the data: {e}')
+
+    #     return df